+++ CHANGELOG.md
@@ -213,6 +213,10 @@
 - Update `Akeneo\Pim\Enrichment\Component\Product\Model\AbstractProduct` to
     - remove the `setFamilyId()` method
     - remove the `$categoryIds` public property and  the `$familyId` and `$groupIds` protected properties
+    - add `isDirty()` and `cleanup()` methods 
+- Remove the `findDatagridViewByAlias()` method in `Oro\Bundle\PimDataGridBundle\Repository\DatagridViewRepositoryInterface`
+- Change constructor of `Akeneo\Pim\Enrichment\Component\Product\Job\DeleteProductsAndProductModelsTasklet` to
+    - add `Akeneo\Tool\Component\Batch\Job\JobRepositoryInterface $jobRepository`
 
 ### CLI commands